refactor: ChatMessage 엔티티 MessageType 필드 추가 / MessageType CLOSE 값 삭제
