+++ back/src/main/java/com/side/anything/back/chat/dto/request/ChatMessageRequest.java
@@ -1,5 +1,6 @@
 package com.side.anything.back.chat.dto.request;
 
+import com.side.anything.back.chat.entity.MessageType;
 import lombok.Getter;
 
 @Getter
